+++ src/file_transfer_client.h
@@ -11,7 +11,7 @@ class FileTransferClient {
 public: 
     FileTransferClient(const std::string& dest_ip, const std::string& dest_port, size_t send_buffer_size = 1400, bool enable_mptcp = true);
 
-    void start_transfer(uint32_t kbytes);  //start transfer kbytes data(fix)
+    void start_transfer(uint64_t kbytes);  //start transfer kbytes data(fix)
 
 private: 
     sockaddr_in _addr;
